Changed union ID to uint32

+++ packets/packets.go
@@ -6,31 +6,31 @@ import (
 )
 
 type Packet struct {
-	ID      uint8
+	ID      uint32
 	Content []byte
 }
 
 type PacketWithConnDTO struct {
-	ID      uint8
+	ID      uint32
 	Content []byte
 	Conn    net.Conn
 }
 
 type PacketDTO struct {
-	ID           uint8
+	ID           uint32
 	Message      interface{}
 	Connection   net.Conn
 	PacketString string
 }
 
 type ReturnPacketDTO struct {
-	ID      uint8
+	ID      uint32
 	Content []byte
 	Conn    net.Conn
 }
 
 type ReturnPacket struct {
-	ID      uint8
+	ID      uint32
 	Content []byte
 }
 
@@ -44,12 +44,12 @@ type TokenPacketContent struct {
 	Signature []byte
 }
 
-var PACKETS_INT_TO_STRING = map[uint8]string{
+var PACKETS_INT_TO_STRING = map[uint32]string{
 	101: "MSG_LOGIN",
 	102: "MSG_REGISTER",
 }
 
-var PACKETS_STRING_TO_INT = map[string]uint8{
+var PACKETS_STRING_TO_INT = map[string]uint32{
 	"MSG_LOGIN":    101,
 	"MSG_REGISTER": 102,
 }
